final changes, just have to debug and fix linting errors now
